[feat] mock api도 수정사항에 맞게 변경
